reorganized file structure and updated import paths
